+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+- Add support for negative matchers
+
 ## [0.1.0] - 2023-04-23
 
 - Initial release